Fake Social sec number FR generator
Out of random number generation

+++ name_DB_gen/social_sec_check_key.py
@@ -4,10 +4,10 @@
 # Description: Script checking french social security identification number, 
 # checking the control key (last 2 digits)
 
-ss =float(
+ss =int(
 	input(
 		"""Please type the social security number, without the control key: """)
 	)
 
-a = int(97-(ss-((round((ss/97.), 0))*97)))
+a = int(97-(ss-((round((ss/97), 0))*97)))
 print("The control key is: ", a)
